all api is ready to sonsume. Everything is ok. Finished.

+++ app/Http/Controllers/MeetingController.php
@@ -5,13 +5,17 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Carbon\Carbon;
 use App\Meeting;
+use JWTAuth;
+
 
 class MeetingController extends Controller
 {
 
     public function __construct()
     {
-        // this-> middleware('name');
+        return $this->middleware('jwt.auth', ['only' => 
+            ['store', 'update', 'destroy']
+            ]);
     }
     /**
      * Display a listing of the resource.
@@ -51,13 +55,17 @@ class MeetingController extends Controller
             'title' => 'required',
             'description' => 'required',
             'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
         ]);
+
+        if (!$user = JWTAuth::parsetoken()->authenticate()) {
+            return response()-> json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        // $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'title' => $title,
@@ -125,14 +133,15 @@ class MeetingController extends Controller
             'time' => 'required|date_format:YmdHie'
         ]);
 
-        // if (! $user = JWTAuth::parseToken()->authenticate()) {
-        //     return response()->json(['msg' => 'User not found'], 404);
-        // }
+        if (! $user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        // $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
 
@@ -168,6 +177,15 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::with('users')->findOrFail($id);
+
+        if (!$user = JWTAuth::parsetoken()->authenticate()) {
+            return response()-> json(['msg' => 'User not found'], 404);
+        }
+
+        if (!$meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
+        };
+
         $users= $meeting->users;
         $meeting->users()->detach();
